Let's break cycle finding to speed it up.
I'm gonna have to find a good way to to this at some future point. Turns
out the problem I'm trying to solve is the "minimum feedback arc set"
problem and is NP-Hard _and_ APX-Hard. I guess there's a reason they
don't teach the directed version of the MST problem in algo classes!