Adding memory management
This is a big one which I couldn't easily break into a few commits.
Sorry.

Anyway, all blocks are now ref counted with boost::intrusive_ptr. When
they are connected in the database, it gets the intrusive_ptr, thus
beginning the refcounting for the block.

Ports which belong to blocks but are dynamically allocated (like with
arrays or ports) are now managed with unique_ptr.

All other objects are -- for the most part -- managed with shared_ptr.