Changing the semantics of registers to match C/C++
Previously, if a register did not contain valid data, a read would block
until data was written. This doesn't match the C semantic of allowing
uninitialized data to be read. It also created extra, unnecessary cycles
in the connection graph. This commit changes the semantic to make
register reads always non-blocking, allowing uninitialized data to be
read.