LLVM Translator: Eliminating the general inputMap
The inputMap for both non-phi inputs and phi inputs just wasn't working
out. PHIs are special and require special treatment. End of story. This
fixed a bug I was tracking down all evening in the sort test. It's also
a clearer design. I shouldn't have used the inputMap in the first place.